requirements.txt
# Core runtime
accelerate==1.3.0
numpy
pandas
Pillow
PyYAML>=6.0
tqdm
torch
safetensors

# Generative pipelines (pin to known working commits)
transformers @ git+https://github.com/huggingface/transformers.git@9d2056f12b66e64978f78a2dcb023f65b2be2108
diffusers @ git+https://github.com/huggingface/diffusers@a4c1aac3ae10172f4acb8eaf83aac7f1f6e02ab0

# GeneVal verifier stack
open-clip-torch
clip-benchmark
mmdet

# Latent verifier stack
llava-hf

# Optional verifiers imported by the script (install only if you use them)
image-reward
t2v-metrics
git+https://github.com/openai/CLIP.git
